adding the employee and its partails

+++ DotNetCorePayroll.Domain/CompanyPayrollSetting.cs
@@ -8,7 +8,7 @@
         public int? HoursPerDay { get; set; }
         public int? WeeklyPeriods { get; set; }
         public int? HourPerWeek { get; set; }
-        public decimal? DaysPerMonth { get; set; }
+        public double? DaysPerMonth { get; set; }
         public string PayrollMessage { get; set; }
 
         public virtual Company Company { get; set; }
